Splitting time_integrator.hpp into several modules.
